refactor: move price formatting into helpers
