+++ packages/client/src/components/manage-users/requested-reserved-table.vue
@@ -1,6 +1,6 @@
 <template>
   <dialog-table
-    :class="isMobile ? 'sticky-last-column' : undefined"
+    :class="isMobile ? 'sticky-last-column' : ''"
     :rows="rows"
     :columns="columns"
   >
@@ -50,7 +50,7 @@ import { QTableColumn, QTableProps } from "quasar";
 import { computed } from "vue";
 import { useI18n } from "vue-i18n";
 import { useLateralDrawer } from "src/composables/use-lateral-drawer";
-import { formatPrice } from "src/composables/use-misc-formats";
+import { formatPrice } from "src/helpers/formatting";
 import { RequestSummaryFragment } from "src/services/request.graphql";
 import { ReservationSummaryFragment } from "src/services/reservation.graphql";
 import UtilityChip from "../utility-chip.vue";
